gemini-agents/requirements.txt
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
google-genai>=1.49.0
firebase-admin